+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Support non-JSON content types in `WriteAttachments` and `ReadAttachments`, [PR-65](https://github.com/reductstore/reduct-go/pull/65)
+
 ## 1.19.1 - 2026-04-22
 
 ### Fixed